+++ apply_cp_patches.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
-# Add audioif's audiodynamics, audioroute, audiomath, audioecho,
-# audioconvolve and audiobiquad modules
+# Add audioif's audiodynamics, audioroute (Splitter, MidSide),
+# audiomath, audioecho, audioconvolve and audiobiquad modules
 # to a CircuitPython tree.
 #
 #   ./apply_cp_patches.sh --dry-run [--port PORT] [--variant VARIANT]
@@ -250,6 +250,7 @@ if [[ "$MODE" == "--status" ]]; then
                 shared-bindings/audioconvolve/__init__.c \
                 shared-bindings/audiobiquad/__init__.c \
                 shared/audioif_dynamics.c shared/audioif_splitter.c \
+                shared/audioif_midside.c \
                 shared/audioif_multiply.c shared/audioif_feedback_delay.c \
                 shared/audioif_trig.c shared/audioif_fft.c \
                 shared/audioif_convolve.c shared/audioif_filter_f32.c; do
@@ -339,6 +340,7 @@ BINDING_ANCHOR=$'\tshared-bindings/audiofilters/__init__.c \\'
 MODULE_ANCHOR=$'\tshared-module/audiofilters/__init__.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audiodynamics/Dynamics.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audiodynamics/__init__.c \\'
+insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audioroute/MidSide.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audioroute/Splitter.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audioroute/SplitterTap.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audioroute/__init__.c \\'
@@ -352,6 +354,7 @@ insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audiobiqua
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audiobiquad/AllPass.c \\'
 insert_line_after "$VARIANT_MK" "$BINDING_ANCHOR" $'\tshared-bindings/audiobiquad/__init__.c \\'
 insert_line_after "$VARIANT_MK" "$MODULE_ANCHOR" $'\tshared-module/audiodynamics/Dynamics.c \\'
+insert_line_after "$VARIANT_MK" "$MODULE_ANCHOR" $'\tshared-module/audioroute/MidSide.c \\'
 insert_line_after "$VARIANT_MK" "$MODULE_ANCHOR" $'\tshared-module/audioroute/Splitter.c \\'
 insert_line_after "$VARIANT_MK" "$MODULE_ANCHOR" $'\tshared-module/audioroute/SplitterTap.c \\'
 insert_line_after "$VARIANT_MK" "$MODULE_ANCHOR" $'\tshared-module/audiomath/Multiply.c \\'